src/main/java/com/example/crudexample/service/ProductService.java
package com.example.crudexample.service;

import com.example.crudexample.entity.Product;
import com.example.crudexample.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class ProductService {

    @Autowired
    private ProductRepository productRepository;

    public List<Product> findAll() {
        return productRepository.findAll();
    }

    public Optional<Product> findById(Long id) {
        if(productRepository.existsById(id)){
            return productRepository.findById(id);
        }

        return Optional.empty();
    }

    public Product save(Product product) {
        return productRepository.save(product);
    }

    public void deleteById(Long id) {
        productRepository.deleteById(id);
    }
}
